Simplify, bring back changes from yesterday.
