--- tests/test_dag_tree/conftest.py
@@ -1,85 +1,0 @@
-"""Test fixtures for dag_tree tests."""
-
-from __future__ import annotations
-
-from typing import TYPE_CHECKING
-
-import pytest
-
-if TYPE_CHECKING:
-    from dags.dag_tree.typing import NestedFunctionDict
-
-
-def _global__f(g, namespace1__f1, input_, namespace1__input):
-    """Global function, duplicate simple name."""
-    return {
-        "name": "global__f",
-        "args": {
-            "g": g,
-            "namespace1__f1": namespace1__f1,
-            "input_": input_,
-            "namespace1__input": namespace1__input,
-        },
-    }
-
-
-def _global__g():
-    """Global function, unique simple name."""
-    return {"name": "global__g"}
-
-
-def _namespace1__f(
-    g,
-    namespace1__f1,
-    namespace2__f2,
-    namespace1__input,
-    namespace2__input2,
-):
-    """Namespaced function, duplicate simple name."""
-    return {
-        "name": "namespace1__f",
-        "args": {
-            "g": g,
-            "namespace1__f1": namespace1__f1,
-            "namespace2__f2": namespace2__f2,
-            "namespace1__input": namespace1__input,
-            "namespace2__input2": namespace2__input2,
-        },
-    }
-
-
-def _namespace1__f1():
-    """Namespaced function, unique simple name."""
-    return {"name": "namespace1__f1"}
-
-
-def _namespace2__f(f2, input_):
-    """Namespaced function, duplicate simple name. All arguments with simple names."""
-    return {"name": "namespace2__f", "args": {"f2": f2, "input_": input_}}
-
-
-def _namespace2__f2():
-    """Namespaced function, unique simple name."""
-    return {"name": "namespace2__f2"}
-
-
-def _namespace1__deep__f():
-    """A deeply nested function."""
-    return {"name": "namespace1_deep__f"}
-
-
-@pytest.fixture
-def functions() -> NestedFunctionDict:
-    return {
-        "f": _global__f,
-        "g": _global__g,
-        "namespace1": {
-            "f": _namespace1__f,
-            "f1": _namespace1__f1,
-            "deep": {"f": _namespace1__deep__f},
-        },
-        "namespace2": {
-            "f": _namespace2__f,
-            "f2": _namespace2__f2,
-        },
-    }
