+++ tests/test_dag_tree/test_parameters.py
@@ -20,13 +20,80 @@ if TYPE_CHECKING:
         NestedInputStructureDict,
     )
 
-# Import fixtures
-from .conftest import (
-    _global__f,
-    _global__g,
-    _namespace1__f,
-    _namespace2__f,
-)
+
+def f(g, namespace1__f1, global_input, namespace1__input):
+    """Global function, duplicate simple name."""
+    return {
+        "name": "f",
+        "args": {
+            "g": g,
+            "namespace1__f1": namespace1__f1,
+            "global_input": global_input,
+            "namespace1__input": namespace1__input,
+        },
+    }
+
+
+def g():
+    """Global function, unique simple name."""
+    return {"name": "g"}
+
+
+def _namespace1__f(
+    g,
+    namespace1__f1,
+    namespace2__f2,
+    namespace1__input,
+    namespace2__input2,
+):
+    """Namespaced function, duplicate simple name."""
+    return {
+        "name": "namespace1__f",
+        "args": {
+            "g": g,
+            "namespace1__f1": namespace1__f1,
+            "namespace2__f2": namespace2__f2,
+            "namespace1__input": namespace1__input,
+            "namespace2__input2": namespace2__input2,
+        },
+    }
+
+
+def _namespace1__f1():
+    """Namespaced function, unique simple name."""
+    return {"name": "namespace1__f1"}
+
+
+def _namespace2__f(f2, global_input):
+    """Namespaced function, duplicate simple name. All arguments with simple names."""
+    return {"name": "namespace2__f", "args": {"f2": f2, "global_input": global_input}}
+
+
+def _namespace2__f2():
+    """Namespaced function, unique simple name."""
+    return {"name": "namespace2__f2"}
+
+
+def _namespace1__deep__f():
+    """A deeply nested function."""
+    return {"name": "namespace1_deep__f"}
+
+
+@pytest.fixture
+def functions() -> NestedFunctionDict:
+    return {
+        "f": f,
+        "g": g,
+        "namespace1": {
+            "f": _namespace1__f,
+            "f1": _namespace1__f1,
+            "deep": {"f": _namespace1__deep__f},
+        },
+        "namespace2": {
+            "f": _namespace2__f,
+            "f2": _namespace2__f2,
+        },
+    }
 
 
 @pytest.mark.parametrize(
@@ -35,20 +102,20 @@ from .conftest import (
         (
             {},
             "",
-            _global__g,
+            g,
             {},
         ),
         (
-            {"input_": None},
+            {"global_input": None},
             "namespace2",
             _namespace2__f,
-            {"f2": "namespace2__f2", "input_": "input_"},
+            {"f2": "namespace2__f2", "global_input": "global_input"},
         ),
         (
-            {"namespace2": {"input_": None}},
+            {"namespace2": {"global_input": None}},
             "namespace2",
             _namespace2__f,
-            {"f2": "namespace2__f2", "input_": "namespace2__input_"},
+            {"f2": "namespace2__f2", "global_input": "namespace2__global_input"},
         ),
         (
             {"namespace1": {"input": None}, "namespace2": {"input2": None}},
@@ -63,13 +130,13 @@ from .conftest import (
             },
         ),
         (
-            {"input_": None, "namespace1": {"input_": None}},
+            {"global_input": None, "namespace1": {"global_input": None}},
             "",
-            _global__f,
+            f,
             {
                 "g": "g",
                 "namespace1__f1": "namespace1__f1",
-                "input_": "input_",
+                "global_input": "global_input",
                 "namespace1__input": "namespace1__input",
             },
         ),
